refactor : challenge 테이블에서 화면을 그리기 위한 용도의 user_count 컬럼을 제거함

+++ src/main/java/com/challengers/userchallenge/repository/UserChallengeRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 
 public interface UserChallengeRepository extends JpaRepository<UserChallenge,Long>, UserChallengeRepositoryCustom {
     Optional<UserChallenge> findByUserIdAndChallengeId(Long userId, Long challengeId);
-    Long countByChallengeId(Long challengeId);
+    int countByChallengeId(Long challengeId);
 
     List<UserChallenge> findByChallengeId(Long challengeId);
